question_analysis_v02/requirements.txt
openai==1.98.0
python-dotenv>=1.0.1
httpx>=0.27.2
tqdm>=4.66.4